web/mirror/ohrnightly.sh: archive all nightlies and svninfo.txt too, not just *.zip files
git-svn-id: https://rpg.hamsterrepublic.com/source/web@8293 7d344553-34f0-0310-a9b1-970ce8f1c3a2

+++ mirror/ohrnightly/mirror.sh
@@ -39,6 +39,13 @@ httrack \
   "${URL}" \
   -"*" \
   +"*.zip" \
+  +"*.gz" \
+  +"*.bz2" \
+  +"*.apk" \
+  +"*.exe" \
+  +"*.dmg" \
+  +"*.deb" \
+  +"*svninfo.txt" \
   -"*-default.zip" \
   > /dev/null
 
@@ -55,29 +62,36 @@ fi
 YEST=`date -d "Yesterday" "+%Y-%m-%d"`
 YDIR="../${YEST}"
 
-for i in *.zip ; do
+for i in *.zip *.bz2 *.gz *.apk *.dmg *.exe *.deb *.txt; do
   OLD="${YDIR}/${i}"
   if [ ! -f "${OLD}" ] ; then
     echo "${i} (NEW)"
     continue
   fi
-  TMP1="/tmp/ohrnightly.${RANDOM}.tmp"
-  TMP2="/tmp/ohrnightly.${RANDOM}.tmp"
-  unzip -qq -d "${TMP1}" "${i}"
-  unzip -qq -d "${TMP2}" "${OLD}"
-  if [ -f "${TMP1}/svninfo.txt" -a -f "${TMP2}/svninfo.txt" ] ; then
-    DIF=`diff -u "${TMP1}/svninfo.txt" "${TMP2}/svninfo.txt"`
+  if [ ${i: -4} == ".zip" ]; then
+    # Extract svninfo.txt from .zip files, most of which are Windows
+    # nightlies, to check whether they have changed.
+    TMP1="/tmp/ohrnightly.${RANDOM}.tmp"
+    TMP2="/tmp/ohrnightly.${RANDOM}.tmp"
+    unzip -qq -d "${TMP1}" "${i}"
+    unzip -qq -d "${TMP2}" "${OLD}"
+    if [ -f "${TMP1}/svninfo.txt" -a -f "${TMP2}/svninfo.txt" ] ; then
+      DIF=`diff -u "${TMP1}/svninfo.txt" "${TMP2}/svninfo.txt"`
+    else
+      DIF=`diff -r "${TMP1}" "${TMP2}"`
+    fi
+    rm -Rf "${TMP1}" "${TMP2}"
   else
-    DIF=`diff -r "${TMP1}" "${TMP2}"`
+    # Mac and Linux nightly builds are only rebuilt if svn changed
+    DIF=`cmp "${i}" "${OLD}"`
   fi
-  rm -Rf "${TMP1}" "${TMP2}"
   if [ "${DIF}" ] ; then
     echo "${i} (Updated)"
   else
     printf "${i}"
     LINK=`followlink "${OLD}"`
     if [ -f "${LINK}" ] ; then
-      rm "${i}"
+      rm -f "${i}"  # why is ohrrpgce-linux-wip-x86.tar.bz2 write-protected?
       ln -s "${LINK}" "${i}"
       echo " (${LINK})"
     else
@@ -85,6 +99,7 @@ for i in *.zip ; do
     fi
   fi
 done
+
 
 # Throw away nightlies older than one year.
 find "${WEBDIR}" -type d -mtime +365 -exec rm -R "{}" \;
